backend-procesamiento/requirements.txt
# Dependencias para el procesador de PDFs para sistema RAG

# Carga de variables de entorno
python-dotenv>=1.0.0

# Procesamiento de PDFs
pdfplumber>=0.10.0

# División de texto para RAG
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Generación de embeddings con OpenAI
openai>=1.0.0

# Base de datos vectorial Pinecone
pinecone>=7.0.0

# Dependencias adicionales
tiktoken>=0.5.0
